% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{continue_MCMC}
\alias{continue_MCMC}
\title{continue_MCMC function}
\usage{
continue_MCMC(
  mcmc,
  iterations,
  moves_per_iteration = 1,
  thinning = 1,
  testing_conditions = NULL
)
}
\arguments{
\item{mcmc}{The ellipseMCMC object to run}

\item{iterations}{The number of iterations to run the analysis. None of these iterations will be used as burnin, and hyperparameters will not be further tuned.}

\item{moves_per_iteration}{The number of moves per iteration (total number of proposals for one iteration). The default value is 1.}

\item{thinning}{How often the mcmc output is recorded, for example (thinning=10) will record every 10 generations. The default value is 1.}

\item{testing_conditions}{A list of extra testing conditions. Options include c("under_prior", "test_probs", "record_proposals"). If no values are provided, the analysis will continue using the preexisting testing conditions.}
}
\description{
Continues an MCMC analysis
}
