% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailedTree.R
\name{detailedTree}
\alias{detailedTree}
\title{detailedTree object}
\description{
This class stores an object of class "phylo" and generates other data relevant to the phylogeny. May take a long time to generate for larger trees.
}
\details{
The "detailedTree" object stores a tree object of class "phylo", typically from the "ape" package, alongside other important information about its nodes, branches, and structure. This information is useful when implementing phylogenetic analyses, like the ellipse evolution analysis.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{tree}}{An object of class "phylo", typically from the "ape" package. This must be an ultrametric, binary tree.}

\item{\code{n_tips}}{An integer giving the number of tips in the tree}

\item{\code{n_nodes}}{An integer giving the number of nodes in the tree, including root and tips}

\item{\code{root_idx}}{An integer giving the index of the root node, typically 1 greater than the number of tips}

\item{\code{edges}}{An object of class "data.frame" listing all edges with start, end, and length}

\item{\code{nodes}}{An object of class "data.frame" listing all nodes with parent, right child, and left child}

\item{\code{descendants}}{An object of class "matrix" indicating whether a node (column) is a descendant of another node (row) using binary representation (0 = non-descendant, 1 = descendant)}

\item{\code{tip_descendants}}{An object of class "matrix" indicating whether a tip (column) is a descendant of another node (row) using binary representation (0 = non-descendant, 1 = descendant), where all tips are descendants of themselves}

\item{\code{phy_matrix}}{An object of class "matrix" giving the phylogenetic variance-covariance matrix indicated by the tree structure (not scaled by evolutionary rates)}

\item{\code{inv_matrix}}{An object of class "matrix" giving the inverse of the phylogenetic variance-covariance matrix indicated by the tree structure (not scaled by evolutionary rates)}

\item{\code{full_matrix}}{An object of class "matrix" giving the full phylogenetic variance-covariance matrix indicated by the tree structure, including internal nodes (not scaled by evolutionary rates)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-detailedTree-new}{\code{detailedTree$new()}}
\item \href{#method-detailedTree-clone}{\code{detailedTree$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-detailedTree-new"></a>}}
\if{latex}{\out{\hypertarget{method-detailedTree-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the detailedTree class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{detailedTree$new(tree)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tree}}{An object of class "phylo", typically from the "ape" package}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-detailedTree-clone"></a>}}
\if{latex}{\out{\hypertarget{method-detailedTree-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{detailedTree$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
