% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{ellipseMCMC}
\alias{ellipseMCMC}
\title{ellipseMCMC object}
\description{
This class stores an object of class "dataTree" and performs an MCMC on the dataTree using a customizeable set of moves.
}
\details{
If the dataTree provided by the user does not contain initialized data-augmented values for nodes (both W & V), initialized values for model parameters (param), or tip states, values will be simulated.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{dataTree}}{An object of class "dataTree"}

\item{\code{filepath}}{An object of class "character" (a string) that gives the filepath where you want the output log files saved. Default behavior will save the output to the same directory where the mcmc is run.}

\item{\code{prefix}}{An object of class "character" (a string) that gives a prefix to all logfiles. Default is no prefix.}

\item{\code{prior_sigma_x}}{Prior on the sigma_x model parameter. Should be an object of type "distribution" generated by using the package "distributions3". Default is Uniform(0,1)}

\item{\code{prior_sigma_y}}{Prior on the sigma_y model parameter. Should be an object of type "distribution" generated by using the package "distributions3". Default is Uniform(0,1)}

\item{\code{prior_sigma_r}}{Prior on the sigma_r model parameter. Should be an object of type "distribution" generated by using the package "distributions3". Default is Uniform(0,1)}

\item{\code{prior_sigma_s}}{Prior on the sigma_s model parameter. Should be an object of type "distribution" generated by using the package "distributions3". Default is Uniform(0,1)}

\item{\code{prior_sigma_a}}{Prior on the sigma_a model parameter. Should be an object of type "distribution" generated by using the package "distributions3". Default is Uniform(0,5)}

\item{\code{prior_root_x}}{Prior on the root_x model parameter. Should be an object of type "distribution" generated by using the package "distributions3". Default is Normal(0,1)}

\item{\code{prior_root_y}}{Prior on the root_y model parameter. Should be an object of type "distribution" generated by using the package "distributions3". Default is Normal(0,1)}

\item{\code{prior_root_r}}{Prior on the root_r model parameter. Should be an object of type "distribution" generated by using the package "distributions3". Default is Normal(0,5)}

\item{\code{prior_root_s}}{Prior on the root_s model parameter. Should be an object of type "distribution" generated by using the package "distributions3". Default is Normal(0,5)}

\item{\code{prior_root_a}}{Prior on the root_a model parameter. Should be an object of type "distribution" generated by using the package "distributions3". Default is Uniform(-5,10)}

\item{\code{prior_mu}}{Prior on the mu model parameter. Should be an object of type "distribution" generated by using the package "distributions3". Default is Uniform(-5,10)}

\item{\code{prior_kappa}}{Prior on the kappa model parameter. Should be an object of the type "distribution" generated by using the package "distributions3". Default is Uniform(0,5)}

\item{\code{prob_W}}{The log probability of the data-augmented history of phylogenetic scenarios under the model}

\item{\code{prob_V}}{The log probability of the data-augmented history of r, s, and a values under the model}

\item{\code{prob_XY}}{The log probability of the observed x and y characters under the model}

\item{\code{prob_x}}{The log probability of the observed x character under the model}

\item{\code{prob_y}}{The log probability of the observed y character under the model}

\item{\code{prob_r}}{The log probability of the data-augmented history of r values under the model}

\item{\code{prob_s}}{The log probability of the data-augmented history of s values under the model}

\item{\code{prob_a}}{The log probability of the data-augmented history of a values under the model}

\item{\code{expected_values}}{The expected values of x and y under the model, based on the data-augmented history}

\item{\code{branch_probs}}{The branch-wise probabilities of r, s, and a, based on the data-augmented history}

\item{\code{likelihood}}{The log likelihood of the model under the priors (includes the data-augmented history)}

\item{\code{prior}}{The log prior of the model}

\item{\code{posterior}}{The log posterior of the model}

\item{\code{acceptances}}{A list of lists that keeps track of the number of proposals of each type and the number of acceptances, alongside the acceptance probabilities and proposal hyperparameters (for tuning during burnin)}

\item{\code{proposal_weights}}{A list of MCMC proposals and their relative weights (can use proportions or whole numbers). If you want to exclude a proposal, the weight can be set to 0}

\item{\code{completed_iterations}}{The total number of iterations this MCMC object has completed}

\item{\code{elapsed_time}}{The total amount of time this MCMC has spent running, in hms format (from hms package)}

\item{\code{current_proposal}}{The proposal currently being performed (or last performed) by the MCMC}

\item{\code{testing_conditions}}{A list of extra testing conditions. Options include c("under_prior", "test_probs", "record_proposals")}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ellipseMCMC-new}{\code{ellipseMCMC$new()}}
\item \href{#method-ellipseMCMC-run}{\code{ellipseMCMC$run()}}
\item \href{#method-ellipseMCMC-continue}{\code{ellipseMCMC$continue()}}
\item \href{#method-ellipseMCMC-set_empirical_priors}{\code{ellipseMCMC$set_empirical_priors()}}
\item \href{#method-ellipseMCMC-clone}{\code{ellipseMCMC$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ellipseMCMC-new"></a>}}
\if{latex}{\out{\hypertarget{method-ellipseMCMC-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the ellipseMCMC class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ellipseMCMC$new(
  dataTree = NULL,
  prior_sigma_x = distributions3::Uniform(0, 1),
  prior_sigma_y = distributions3::Uniform(0, 1),
  prior_sigma_r = distributions3::Uniform(0, 1),
  prior_sigma_s = distributions3::Uniform(0, 1),
  prior_sigma_a = distributions3::Uniform(0, 5),
  prior_root_x = distributions3::Normal(0, 1),
  prior_root_y = distributions3::Normal(0, 1),
  prior_root_r = distributions3::Normal(0, 5),
  prior_root_s = distributions3::Normal(0, 5),
  prior_root_a = distributions3::Uniform(-5, 10),
  prior_mu = distributions3::Uniform(-5, 10),
  prior_kappa = distributions3::Uniform(0, 5),
  proposal_weights = list(sigma_x = 5, sigma_y = 5, sigma_r = 5, sigma_s = 5, sigma_a =
    5, root_x = 3, root_y = 3, root_r = 3, root_s = 3, root_a = 3, mu = 5, kappa = 5, W_d
    = 1.5, W_m = 1.5, W_c = 1.5, W_h = 1.5, V_r = 4, V_s = 4, V_a = 4, tip = 1.5)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataTree}}{An object of class "dataTree"}

\item{\code{prior_sigma_x}}{Prior on the sigma_x model parameter. Should be an object of type "distribution" generated by using the package "distributions3". Default is Uniform(0,1)}

\item{\code{prior_sigma_y}}{Prior on the sigma_y model parameter. Should be an object of type "distribution" generated by using the package "distributions3". Default is Uniform(0,1)}

\item{\code{prior_sigma_r}}{Prior on the sigma_r model parameter. Should be an object of type "distribution" generated by using the package "distributions3". Default is Uniform(0,1)}

\item{\code{prior_sigma_s}}{Prior on the sigma_s model parameter. Should be an object of type "distribution" generated by using the package "distributions3". Default is Uniform(0,1)}

\item{\code{prior_sigma_a}}{Prior on the sigma_a model parameter. Should be an object of type "distribution" generated by using the package "distributions3". Default is Uniform(0,5)}

\item{\code{prior_root_x}}{Prior on the root_x model parameter. Should be an object of type "distribution" generated by using the package "distributions3". Default is Normal(0,1)}

\item{\code{prior_root_y}}{Prior on the root_y model parameter. Should be an object of type "distribution" generated by using the package "distributions3". Default is Normal(0,1)}

\item{\code{prior_root_r}}{Prior on the root_r model parameter. Should be an object of type "distribution" generated by using the package "distributions3". Default is Normal(0,5)}

\item{\code{prior_root_s}}{Prior on the root_s model parameter. Should be an object of type "distribution" generated by using the package "distributions3". Default is Normal(0,5)}

\item{\code{prior_root_a}}{Prior on the root_a model parameter. Should be an object of type "distribution" generated by using the package "distributions3". Default is Uniform(-5,10)}

\item{\code{prior_mu}}{Prior on the mu model parameter. Should be an object of type "distribution" generated by using the package "distributions3". Default is Uniform(-5,10)}

\item{\code{prior_kappa}}{Prior on the kappa model parameter. Should be an object of type "distribution" generated by using the package "distributions3". Default is Uniform(0,5)}

\item{\code{proposal_weights}}{A list of MCMC proposals and their relative weights (can use proportions or whole numbers). If you want to exclude a proposal, the weight can be set to 0}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ellipseMCMC-run"></a>}}
\if{latex}{\out{\hypertarget{method-ellipseMCMC-run}{}}}
\subsection{Method \code{run()}}{
Runs an MCMC analysis
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ellipseMCMC$run(
  iterations,
  moves_per_iteration = 1,
  burnin = 0,
  thinning = 1,
  testing_conditions = c("none"),
  filepath = ".",
  prefix = ""
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{iterations}}{The number of iterations to run the analysis. If a number of burnin generations is also provided, some of these iterations will be burnin (recorded, but should be discarded due to tuning proposal hyperparameters)}

\item{\code{moves_per_iteration}}{The number of moves per iteration (total number of proposals for one iteration). The default value is 1.}

\item{\code{burnin}}{The number of generations that should be used for tuning burnin. These generations will still be saved. The default value is 0.}

\item{\code{thinning}}{How often the mcmc output is recorded, for example (thinning=10) will record every 10 generations. The default value is 1.}

\item{\code{testing_conditions}}{A list of extra testing conditions. Options include c("under_prior", "test_probs", "record_proposals")}

\item{\code{filepath}}{An object of class "character" (a string) that gives the filepath where you want the output log files saved. Default behavior will save the output to the same directory where the mcmc is run.}

\item{\code{prefix}}{An object of class "character" (a string) that gives a prefix to all logfiles. Default is no prefix.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ellipseMCMC-continue"></a>}}
\if{latex}{\out{\hypertarget{method-ellipseMCMC-continue}{}}}
\subsection{Method \code{continue()}}{
Continues an MCMC analysis
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ellipseMCMC$continue(
  iterations,
  moves_per_iteration = 1,
  thinning = 1,
  testing_conditions = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{iterations}}{The number of iterations to run the analysis. None of these iterations will be used as burnin, and hyperparameters will not be further tuned.}

\item{\code{moves_per_iteration}}{The number of moves per iteration (total number of proposals for one iteration). The default value is 1.}

\item{\code{thinning}}{How often the mcmc output is recorded, for example (thinning=10) will record every 10 generations. The default value is 1.}

\item{\code{testing_conditions}}{A list of extra testing conditions. Options include c("under_prior", "test_probs", "record_proposals"). If no values are provided, the analysis will continue using the preexisting testing conditions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ellipseMCMC-set_empirical_priors"></a>}}
\if{latex}{\out{\hypertarget{method-ellipseMCMC-set_empirical_priors}{}}}
\subsection{Method \code{set_empirical_priors()}}{
Sets priors based on empirical data, recommended if you do not intend to define your own priors. Empirical root and mu priors are normal distributions centered on the empirical mean with standard deviations equal to the range of the data. Empirical sigma priors and kappa are uniform distributions from 0 to the range of the data, except sigma_a, which is a uniform distribution from 0 to 4 times the range of the data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ellipseMCMC$set_empirical_priors()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ellipseMCMC-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ellipseMCMC-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ellipseMCMC$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
