% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipseParam.R
\name{ellipseParam}
\alias{ellipseParam}
\title{ellipseParam object}
\description{
This class stores model parameters for the ellipse evolution model
}
\details{
This object is used during simulation to contain the model parameters used for generating data (W, V, XY). It is also used during analysis to store current parameter values.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{cval}}{An object of class "vector" containing any number of numeric elements indicating the values associated with different concentric circles relative to the size of the original circle. The default value includes 4 concentric circles: 0, 0.5, 1.0, and 1.5.}

\item{\code{hval}}{An object of class "vector" containing any number of numeric elements indicating the values associated with different direction lines in radians. The default value includes 8 direction lines: 0, pi/4, pi/2, 3pi/4, pi, 5pi/4, 3pi/2, and 7pi/4.}

\item{\code{rho_m}}{An object of class "vector" containing two numeric elements: the prior probability of cladogenetic mode 1 (budding) and the prior probability of cladogenetic mode 2 (splitting). The default value assigns probability 0.5 to both modes.}

\item{\code{rho_c}}{An object of class "vector" containing any number of numeric elements (equal to the length of cval) corresponding to the prior probabilities of different concentric circle assignments. The default value assigns probability 0.25 to each of 4 concentric circles.}

\item{\code{rho_h}}{An object of class "vector" containing any number of numeric elements (equal to the length of hval) corresponding to the prior probabilities of different direction line assignments. The default value assigns probability 0.125 to each of 8 direction lines.}

\item{\code{sigma_x}}{An object of class "numeric" containing the evolutionary rate of the continuous character 'x', which helps define an ellipse's center location. If left blank, the default behavior will draw a value from a uniform distribution from 0 to 1.}

\item{\code{sigma_y}}{An object of class "numeric" containing the evolutionary rate of the continuous character 'y', which helps define an ellipse's center location. If left blank, the default behavior will draw a value from a uniform distribution from 0 to 1.}

\item{\code{sigma_r}}{An object of class "numeric" containing the evolutionary rate of the continuous character 'r', which helps define an ellipse's "oblongness". If left blank, the default behavior will draw a value from a uniform distribution from 0 to 1.}

\item{\code{sigma_s}}{An object of class "numeric" containing the evolutionary rate of the continuous character 's', which helps define an ellipse's "oblongness". If left blank, the default behavior will draw a value from a uniform distribution from 0 to 1.}

\item{\code{sigma_a}}{An object of class "numeric" containing the evolutionary rate of the continuous character 'a', which helps define an ellipse's size. If left blank, the default behavior will draw a value from a uniform distribution from 0 to 5.}

\item{\code{root_x}}{An object of class "numeric" containing the root value for the 'x' character. If left blank, the default behavior will draw a value from a normal distribution with mean 0 and variance 1.}

\item{\code{root_y}}{An object of class "numeric" containing the root value for the 'y' character. If left blank, the default behavior will draw a value from a normal distribution with mean 0 and variance 1.}

\item{\code{root_r}}{An object of class "numeric" containing the root value for the 'r' character. If left blank, the default behavior will draw a value from a normal distribution with mean 0 and variance 5.}

\item{\code{root_s}}{An object of class "numeric" containing the root value for the 's' character. If left blank, the default behavior will draw a value from a normal distribution with mean 0 and variance 5.}

\item{\code{root_a}}{An object of class "numeric" containing the root value for the 'a' character. If left blank, the default behavior will draw a value from a uniform distribution from -5 to 10.}

\item{\code{mu}}{An object of class "numeric" containing the central tendency of the continuous character 'a', which helps define an ellipse's size. If left blank, the default behavior will draw a value from a uniform distribution from -5 to 10.}

\item{\code{kappa}}{An object of class "numeric" containing the rate of mean reversion of the continuous character 'a', which helps define an ellipse's size. If left blank, the default behavior will draw a value from a uniform distribution from 0 to 5.}

\item{\code{alpha}}{An object of class "numeric" containing the log size of the new ellipse (daughter 2) under the sympatric cladogenetic scenario. Default value is -5.}

\item{\code{z}}{An object of class "numeric" containing the height of the tilt vector for the ellipses. This model parameter does not evolve over time, and is only used in the interpretation of rates of 'r' and 's'. The default value will suffice for most analyses. The default value is 10.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ellipseParam-new}{\code{ellipseParam$new()}}
\item \href{#method-ellipseParam-clone}{\code{ellipseParam$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ellipseParam-new"></a>}}
\if{latex}{\out{\hypertarget{method-ellipseParam-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of the ellipseParam class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ellipseParam$new(
  cval = NULL,
  hval = NULL,
  rho_m = NULL,
  rho_c = NULL,
  rho_h = NULL,
  sigma_r = NULL,
  sigma_s = NULL,
  sigma_a = NULL,
  alpha = NULL,
  root_r = NULL,
  root_s = NULL,
  root_a = NULL,
  sigma_x = NULL,
  sigma_y = NULL,
  root_x = NULL,
  root_y = NULL,
  z = NULL,
  mu = NULL,
  kappa = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cval}}{An object of class "vector" containing any number of numeric elements indicating the values associated with different concentric circles relative to the size of the original circle. The default value includes 4 concentric circles: 0, 0.5, 1.0, and 1.5.}

\item{\code{hval}}{An object of class "vector" containing any number of numeric elements indicating the values associated with different direction lines in radians. The default value includes 8 direction lines: 0, pi/4, pi/2, 3pi/4, pi, 5pi/4, 3pi/2, and 7pi/4.}

\item{\code{rho_m}}{An object of class "vector" containing two numeric elements: the prior probability of cladogenetic mode 1 (budding) and the prior probability of cladogenetic mode 2 (splitting). The default value assigns probability 0.5 to both modes.}

\item{\code{rho_c}}{An object of class "vector" containing any number of numeric elements (equal to the length of cval) corresponding to the prior probabilities of different concentric circle assignments. The default value assigns probability 0.25 to each of 4 concentric circles.}

\item{\code{rho_h}}{An object of class "vector" containing any number of numeric elements (equal to the length of hval) corresponding to the prior probabilities of different direction line assignments. The default value assigns probability 0.125 to each of 8 direction lines.}

\item{\code{sigma_r}}{An object of class "numeric" containing the evolutionary rate of the continuous character 'r', which helps define an ellipse's "oblongness". If left blank, the default behavior will draw a value from a uniform distribution from 0 to 1.}

\item{\code{sigma_s}}{An object of class "numeric" containing the evolutionary rate of the continuous character 's', which helps define an ellipse's "oblongness". If left blank, the default behavior will draw a value from a uniform distribution from 0 to 1.}

\item{\code{sigma_a}}{An object of class "numeric" containing the evolutionary rate of the continuous character 'a', which helps define an ellipse's size. If left blank, the default behavior will draw a value from a uniform distribution from 0 to 5.}

\item{\code{alpha}}{An object of class "numeric" containing the log size of the new ellipse (daughter 2) under the sympatric cladogenetic scenario. Default value is -5.}

\item{\code{root_r}}{An object of class "numeric" containing the root value for the 'r' character. If left blank, the default behavior will draw a value from a normal distribution with mean 0 and variance 5.}

\item{\code{root_s}}{An object of class "numeric" containing the root value for the 's' character. If left blank, the default behavior will draw a value from a normal distribution with mean 0 and variance 5.}

\item{\code{root_a}}{An object of class "numeric" containing the root value for the 'a' character. If left blank, the default behavior will draw a value from a normal distribution with mean 0 and variance 5.}

\item{\code{sigma_x}}{An object of class "numeric" containing the evolutionary rate of the continuous character 'x', which helps define an ellipse's center location. If left blank, the default behavior will draw a value from a uniform distribution from 0 to 1.}

\item{\code{sigma_y}}{An object of class "numeric" containing the evolutionary rate of the continuous character 'y', which helps define an ellipse's center location. If left blank, the default behavior will draw a value from a uniform distribution from 0 to 1.}

\item{\code{root_x}}{An object of class "numeric" containing the root value for the 'x' character. If left blank, the default behavior will draw a value from a normal distribution with mean 0 and variance 1.}

\item{\code{root_y}}{An object of class "numeric" containing the root value for the 'y' character. If left blank, the default behavior will draw a value from a normal distribution with mean 0 and variance 1.}

\item{\code{z}}{An object of class "numeric" containing the height of the tilt vector for the ellipses. This model parameter does not evolve over time, and is only used in the interpretation of rates of 'r' and 's'. The default value will suffice for most analyses. The default value is 10.}

\item{\code{mu}}{An object of class "numeric" containing the central tendency of the continuous character 'a', which helps define an ellipse's size. If left blank, the default behavior will draw a value from a normal distribution with mean 0 and variance 5.}

\item{\code{kappa}}{An object of class "numeric" containing the rate of mean reversion of the continuous character 'a', which helps define an ellipse's size. If left blank, the default behavior will draw a value from a uniform distribution from 0 to 5.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ellipseParam-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ellipseParam-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ellipseParam$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
