% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empire-package.R
\docType{package}
\name{empire-package}
\alias{empire}
\alias{empire-package}
\title{empire}
\description{
This package contains functions for simulation and inference under the EMPIRE model. For more information on EMPIRE, see the associated paper.
}
\details{
Roxygen created this manual page on 2026-04-22 using R version
4.3.1.

The EMPIRE package is written in an \emph{object-oriented} manner. Data is stored in objects with custom classes, and those objects have "methods" (functions) that can be applied. For example, you can use \code{dataTree$new()} to create a new dataTree object, or \code{mcmc$run()} to run your mcmc object.

However, there are also a number of "wrapper" functions provided for your convenience. These wrapper functions allow you to interact with EMPIRE in a more conventional "functional" way. The tutorial examples use these wrapper functions.
}
\section{Important Objects}{
\itemize{
\item \code{\link[=ellipseParam]{ellipseParam()}} stores model parameter values
\item \code{\link[=detailedTree]{detailedTree()}} stores a phylogenetic tree and associated information
\item \code{\link[=dataTree]{dataTree()}} stores a detailedTree and any simulated / attached data
\item \code{\link[=ellipseMCMC]{ellipseMCMC()}} stores a dataTree object and necessary values / methods for performing MCMC
}
}

\section{Examples}{
\subsection{SIMULATION EXAMPLE}{

An example which simulates a tree with data and tip states.

Functions: \code{\link[=make_dataTree]{make_dataTree()}}, \code{\link[=sim_data]{sim_data()}}, \code{\link[=save_dataTree]{save_dataTree()}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{# 1. Set directory for output files (modify for your file system)
outpath <- "./output"

# 2. Simulate an example tree using phytools
set.seed(4)
example_tree <- phytools::pbtree(b=.05, d=0, t=40, type="continuous")

# 3. Make this "phylo" tree into a dataTree (an EMPIRE data type)
tree <- make_dataTree(tree=example_tree)

# 4. Simulate some data for this tree
tree <- sim_data(tree=tree)

# 5. Save the tree and tip data
save_dataTree(tree=tree, filepath=outpath, prefix="true")

# 6. Plot the tree and node labels
plot_label_tree(tree=tree, filepath=outpath)

# 7. Plot the grid system for the simulated cladogenetic scenarios
plot_grid(model=tree)
}\if{html}{\out{</div>}}
}

\subsection{ANALYSIS EXAMPLE}{

An example which performs MCMC on the previously-simulated dataset. You will need:
\itemize{
\item A bifurcating, ultrametric phylogeny of type \strong{phylo} (i.e. read in with the "ape" package)
\item A \strong{data.frame} with tip ellipse data; headers should be \emph{taxon}, \emph{x}, \emph{y}, \emph{r}, \emph{s}, \emph{a}.
}

Make sure your taxon names match the tip labels of the phylogeny (can be in any order).

Functions: \code{\link[=make_dataTree]{make_dataTree()}}, \code{\link[=make_MCMC]{make_MCMC()}}, \code{\link[=set_empirical_priors_MCMC]{set_empirical_priors_MCMC()}}, \code{\link[=run_MCMC]{run_MCMC()}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{# 1. Read tree and tip data from directory
outpath <- "./output"
example_tree <- ape::read.tree(paste0(outpath,"/true.tree.txt"))
data <- read.csv(file=paste0(outpath,"/true.tip_data.tsv"), sep="\\t", header=TRUE)

# 2. Turn tree and data into a dataTree (an EMPIRE data type)
tree <- make_dataTree(tree=example_tree, tip_data=data)

# 3. Choose proposal weights (optional)
proposal_weights=list(sigma_x=5,
                      sigma_y=5,
                      sigma_r=5,
                      sigma_s=5,
                      sigma_a=5,
                      root_x=3,
                      root_y=3,
                      root_r=3,
                      root_s=3,
                      root_a=3,
                      mu=5,
                      kappa=5,
                      W_d=1.5,
                      W_m=1.5,
                      W_c=1.5,
                      W_h=1.5,
                      V_r=4,
                      V_s=4,
                      V_a=4,
                      tip=1.5)

# 4. Make MCMC object
mcmc <- make_MCMC(dataTree=tree, proposal_weights=proposal_weights)

# 5. Set priors for MCMC based on the range of the data
mcmc <- set_empirical_priors_MCMC(mcmc)

# 6. Run MCMC
run_MCMC(mcmc=mcmc, iterations=1000, burnin=100, thinning=10, filepath=outpath)
}\if{html}{\out{</div>}}
}

\subsection{POST-PROCESSING EXAMPLE}{

An example for processing the output of an EMPIRE MCMC analysis.

Functions: \code{\link[=reconstruct_ellipses]{reconstruct_ellipses()}}, \code{\link[=plot_scenario]{plot_scenario()}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{# 1. Read tree and tip data from directory
outpath <- "./output"
example_tree <- ape::read.tree(paste0(outpath,"/true.tree.txt"))
data <- read.csv(file=paste0(outpath,"/true.tip_data.tsv"), sep="\\t", header=TRUE)
tree <- make_dataTree(tree=example_tree, tip_data=data)

# 2. Generate reconstructions for each internal node
reconstruction <- reconstruct_ellipses(outpath,burnin=10)
print(reconstruction)
# Note that it includes both indices and values for concentric circles (c, cval) and direction lines (h, hval)
# The VALUES are used for plotting ellipses (e.g. plot_scenario()) or making daughters (transform_node())

# 3. Plot the phylogenetic tree on a map, with nodes located at their reconstructed centroids
plot_tree_map(tree,reconstruction)

# 4. Plot the scenario at the MRCA -- note that the range of the blue (left) daughter is very small
mrca <- reconstruction[1,]
plot_scenario_annotated(x=mrca$x,y=mrca$y,r=mrca$r,s=mrca$s,a=mrca$a,d=mrca$d,m=mrca$m,cval=mrca$cval,hval=mrca$hval,z=10,alpha=-5)

# 5. Plot the scenario at a different node (e.g. node 28)
node <- reconstruction[which(reconstruction$node==28),]
plot_scenario_annotated(x=node$x,y=node$y,r=node$r,s=node$s,a=node$a,d=node$d,m=node$m,cval=node$cval,hval=node$hval,z=10,alpha=-5)

# 6. Plot a lineage, including cladogenetic scenarios, from the MRCA to a specific tip
plot_lineage(tree,reconstruction,1)

# 7. Plot the posterior density for the ancestral and daughter ellipses alongside the best reconstruction at a set of nodes
plot_posterior(outpath,c(21,23))
}\if{html}{\out{</div>}}
}
}

\references{
stuff
}
\author{
Sarah Swiston, \email{sarah.k.swiston@gmail.com}
}
\keyword{internal}
