% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ellipse.R
\name{get_ellipses}
\alias{get_ellipses}
\title{get_ellipses() function}
\usage{
get_ellipses(data = NULL, z = 10, confidence = 0.95)
}
\arguments{
\item{data}{The spatial MULTIPOLYGON to turn into ellipses. Alternatively, the occurrence points to turn into ellipses, as a MULTIPOINT object. The first column of this sf object must contain the taxon names. The column containing data (polygons or points) must be called "geometry".}

\item{z}{The height of the tilt vector for generating the ellipses, default value is 10}

\item{confidence}{The level of confidence used to generate the confidence ellipse, default value is 0.95}
}
\description{
Takes a spatial MULTIPOLYGON (sf package) with entries for each species and generates ellipse parameters x, y, r, s, and a for each species. Returns an object of class "data.frame" (taxon,x,y,r,s,a,z). This is done by sampling points from within each spatial polygon/multipolygon, then creating a confidence ellipse around those points using the stats function cov.wt(). Alternatively, can provide a set of occurrence points in a MULTIPOINT (sf) object.
}
