% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prob.R
\name{get_prob}
\alias{get_prob}
\title{get_prob() function}
\usage{
get_prob(dataTree = NULL, prob_w = NULL, prob_v = NULL, prob_xy = NULL)
}
\arguments{
\item{dataTree}{A dataTree object containing a detailedTree object, and ellipseParam object, and dataframes V, W, and XY. If a probability is provided for a particular set of values (W, V, or XY), the dataFrame(s) required to calculate this probability may be missing from the dataTree without consequence.}

\item{prob_w}{Optionally, the log likelihood of observing the data-augmented cladogenetic scenarios at nodes}

\item{prob_v}{Optionally, the log likelihood of observing the data-augmented values at nodes}

\item{prob_xy}{Optionally, the log likelihood of observing the x and y values at the tips based on the root values, tree structure, and data-augmented values}
}
\description{
Accepts a dataTree object and returns the log likelihood of observing the data augmented history (W,V) and tip values (x and y). Optionally, any of these probabilities can be provided (prob_w, prob_v, prob_xy) and will not be re-computed. If all probabilities are provided, a dataTree is not required.
}
