% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prob.R
\name{get_prob_v}
\alias{get_prob_v}
\title{get_prob_v() function}
\usage{
get_prob_v(dataTree)
}
\arguments{
\item{dataTree}{A dataTree object containing a detailedTree object, and ellipseParam object, and V, a dataframe of data-augmented values (r,s,a) at nodes}
}
\description{
Accepts a dataTree object and returns the log likelihood of observing the data-augmented values V at nodes based on the history of cladogenetic scenarios W
}
