% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prob.R
\name{get_prob_w}
\alias{get_prob_w}
\title{get_prob_w() function}
\usage{
get_prob_w(dataTree)
}
\arguments{
\item{dataTree}{A dataTree object containing a detailedTree object, an ellipseParam object, and W, a dataframe of data-augmented values for cladogenetic scenarios at nodes}
}
\description{
Accepts a dataTree object and returns the log likelihood of observing the data-augmented cladogenetic scenario W
}
