% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prob.R
\name{get_prob_xy}
\alias{get_prob_xy}
\title{get_prob_xy() function}
\usage{
get_prob_xy(dataTree)
}
\arguments{
\item{dataTree}{A dataTree object containing a detailedTree object, an ellipseParam object, and dataframes XY, V, and W containing values for cladogenetic scenarios at nodes}
}
\description{
Accepts a dataTree object and returns the log likelihood of observing the x and y values at the tips based on the root values, tree structure, and data-augmented values
}
