% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTree.R
\name{make_dataTree}
\alias{make_dataTree}
\title{make_dataTree function}
\usage{
make_dataTree(tree, param = NULL, tip_data = NULL)
}
\arguments{
\item{tree}{An object of class "detailedTree", or an object of class "phylo" (typically from the "ape" package)}
}
\description{
Takes an object of class "detailedTree" or "phylo" and creates an object of class "dataTree". May take a long time to generate for larger trees.
}
\details{
The dataTree object is used during simulation to contain the simulated history of all cladogenetic events (W) and continuous values (V, XY) at all nodes (including tips). When initialized, the data fields W, V, and XY will be NULL, and the tree field will be an object of class "detailedTree".
}
