% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailedTree.R
\name{make_detailedTree}
\alias{make_detailedTree}
\title{make_detailedTree() function}
\usage{
make_detailedTree(tree)
}
\arguments{
\item{tree}{An object of class "phylo". This must be an ultrametric, binary tree.}
}
\description{
Turns a tree of class "phylo" into a "detailedTree" object -- this function is a wrapper for detailedTree$new(). May take a long time to generate for larger trees.
}
\details{
The "detailedTree" object stores a tree object of class "phylo", typically from the "ape" package, alongside other important information about its nodes, branches, and structure. This information is useful when implementing phylogenetic analyses, like the ellipse evolution analysis.
}
