% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ellipse.R
\name{make_ellipse_coords}
\alias{make_ellipse_coords}
\title{make_ellipse_coords() function}
\usage{
make_ellipse_coords(x, y, r, s, a, z = 10)
}
\arguments{
\item{x}{The x coordinate of the ellipse centroid}

\item{y}{The y coordinate of the ellipse centroid}

\item{r}{The y coordinate of the ellipse tilt point (when combined with r and z, defines oblongness)}

\item{z}{The height of the tilt vector (when combined with r and s, defines oblongness), default value is 10}
}
\description{
Takes a set of ellipse parameters (x,y,r,s,a,z) and returns a dataframe of points defining the perimeter of the ellipse anticlockwise.
}
