% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_grid}
\alias{plot_grid}
\title{plot_grid() function}
\usage{
plot_grid(model)
}
\arguments{
\item{model}{An object of class ellipseParam, dataTree, or ellipseMCMC with a grid system to plot.}
}
\description{
Plots the grid used for discrete cladogenetic scenarios, returning a ggplot object.
}
