% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_label_tree}
\alias{plot_label_tree}
\title{plot_label_tree() function}
\usage{
plot_label_tree(tree, type = "dated", filepath = ".")
}
\arguments{
\item{tree}{The object of class phylo (ape package), detailedTree, or dataTree to plot}

\item{type}{The type of tree to plot. Options are "dated" or "cladogram", default is "dated"}

\item{filepath}{An object of class "character" giving the filepath to the directory where you want the output to be saved. The default behavior will save files to the current directory.}
}
\description{
Plots a tree including taxon names and node numbers for easy reference. The function returns a ggplot object, but also saves the tree to a file, which is more legible.
}
