% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_lineage}
\alias{plot_lineage}
\title{plot_lineage() function}
\usage{
plot_lineage(
  tree,
  reconstruction,
  tip,
  map = NULL,
  label = FALSE,
  z = 10,
  alpha = -5
)
}
\arguments{
\item{tree}{a \strong{dataTree} object containing your phylogeny and tip data}

\item{reconstruction}{A \strong{data.frame} of ancestral ellipse reconstructions generated by \code{reconstruct_ellipses()}}

\item{tip}{A \strong{character} (string) taxon name or \strong{numeric} tip number of the lineage to plot}

\item{map}{An \strong{sf} polygon to plot in the background (optional)}

\item{label}{Whether to include node and tip numbers, default = FALSE}

\item{z}{The height of the tilt vector, default value is 10}

\item{alpha}{The alpha parameter, default value is -5}
}
\description{
Plots a single evolving lineage on a map, starting with the root of the clade and ending in a particular tip, including reconstructed ancestral ellipses and cladogenetic scenarios.
}
