% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_posterior}
\alias{plot_posterior}
\title{plot_posterior() function}
\usage{
plot_posterior(
  filepath = ".",
  nodes,
  map = NULL,
  burnin = 0,
  thinning = 1,
  cval = NULL,
  hval = NULL,
  z = 10,
  alpha = -5
)
}
\arguments{
\item{filepath}{An object of class "character" (a string) that gives the filepath where the output log files from your MCMC analysis are saved. Default behavior will look for the files in the current directory.}

\item{nodes}{A \strong{numeric} node number or vector node numbers representing the phylogenetic node(s) to plot}

\item{map}{An \strong{sf} polygon to plot in the background (optional)}

\item{burnin}{The number of burnin \emph{rows} produced during the MCMC analysis (burnin divided by thinning); will be removed before processing}

\item{thinning}{Retain every N iterations for plotting the posterior. The default value is 1, retaining all iterations; this may be prohibitively slow with long MCMC chains}

\item{cval}{An object of class "vector" containing any number of numeric elements indicating the values associated with different concentric circles relative to the size of the original circle. The default value includes 4 concentric circles: 0, 0.5, 1.0, and 1.5.}

\item{hval}{An object of class "vector" containing any number of numeric elements indicating the values associated with different direction lines in radians. The default value includes 8 direction lines: 0, pi/4, pi/2, 3pi/4, pi, 5pi/4, 3pi/2, and 7pi/4.}

\item{z}{The height of the tilt vector, default value is 10}

\item{alpha}{The alpha parameter, default value is -5}
}
\description{
Processes output of an ellipse MCMC and generates a posterior distribution for the ancestral ellipse and both daughter ellipses at a single node
}
