% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_tree_map}
\alias{plot_tree_map}
\title{plot_tree_map() function}
\usage{
plot_tree_map(tree, reconstruction, map = NULL, label = FALSE)
}
\arguments{
\item{tree}{a \strong{dataTree} object containing your phylogeny and tip data}

\item{reconstruction}{A \strong{data.frame} of ancestral ellipse reconstructions generated by \code{reconstruct_ellipses()}}

\item{map}{An \strong{sf} polygon to plot in the background (optional)}

\item{label}{Whether to include node and tip numbers, default = FALSE}
}
\description{
Plots the phylogenetic tree on a map, with nodes located at their reconstructed centroids
}
