% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{reconstruct_ellipses}
\alias{reconstruct_ellipses}
\title{reconstruct_ellipses() function}
\usage{
reconstruct_ellipses(filepath = ".", burnin = 0, cval = NULL, hval = NULL)
}
\arguments{
\item{filepath}{An object of class "character" (a string) that gives the filepath where the output log files from your MCMC analysis are saved. Default behavior will look for the files in the current directory.}

\item{burnin}{The number of burnin \emph{rows} produced during the MCMC analysis (burnin divided by thinning); will be removed before processing}

\item{cval}{An object of class "vector" containing any number of numeric elements indicating the values associated with different concentric circles relative to the size of the original circle. The default value includes 4 concentric circles: 0, 0.5, 1.0, and 1.5.}

\item{hval}{An object of class "vector" containing any number of numeric elements indicating the values associated with different direction lines in radians. The default value includes 8 direction lines: 0, pi/4, pi/2, 3pi/4, pi, 5pi/4, 3pi/2, and 7pi/4.}
}
\description{
Processes output of an ellipse MCMC and generates ancestral ellipses for each node
}
