% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{run_MCMC}
\alias{run_MCMC}
\title{run_MCMC function}
\usage{
run_MCMC(
  mcmc,
  iterations,
  moves_per_iteration = 1,
  burnin = 0,
  thinning = 1,
  under_prior = NULL,
  testing_conditions = c("none"),
  filepath = ".",
  prefix = ""
)
}
\arguments{
\item{mcmc}{The ellipseMCMC object to run}

\item{iterations}{The number of iterations to run the analysis. If a number of burnin generations is also provided, some of these iterations will be burnin (recorded, but should be discarded due to tuning proposal hyperparameters)}

\item{moves_per_iteration}{The number of moves per iteration (total number of proposals for one iteration). The default value is 1.}

\item{burnin}{The number of generations that should be used for tuning burnin. These generations will still be saved.}

\item{thinning}{How often the mcmc output is recorded, for example (thinning=10) will record every 10 generations. The default value is 1.}

\item{testing_conditions}{A list of extra testing conditions. Options include c("under_prior", "test_probs", "record_proposals")}

\item{filepath}{An object of class "character" (a string) that gives the filepath where you want the output log files saved. Default behavior will save the output to the same directory where the mcmc is run.}
}
\description{
Runs an MCMC analysis
}
\section{Fields}{

\describe{
\item{\code{prefix}}{An object of class "character" (a string) that gives a prefix to all logfiles. Default is no prefix.}
}}

