% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{set_empirical_priors_MCMC}
\alias{set_empirical_priors_MCMC}
\title{set_empirical_priors_MCMC function}
\usage{
set_empirical_priors_MCMC(mcmc)
}
\arguments{
\item{mcmc}{The ellipseMCMC object for which to set empirical priors, requires tip data}
}
\description{
Sets priors based on empirical data, recommended if you do not intend to define your own priors. Empirical root and mu priors are normal distributions centered on the empirical mean with standard deviations equal to the range of the data. Empirical sigma priors and kappa are uniform distributions from 0 to the range of the data, except sigma_a, which is a uniform distribution from 0 to 4 times the range of the data. Returns the mcmc object with priors set.
}
