% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTree.R
\name{sim_data}
\alias{sim_data}
\title{sim_data function}
\usage{
sim_data(tree, param = NULL)
}
\arguments{
\item{tree}{An object of class "dataTree" containing a "detailedTree" in the tree field, an object of class "detailedTree", or an object of class "phylo" (typically from the "ape" package)}

\item{param}{An object of type ellipseParam containing information about the model parameters associated with a simulation or MCMC iteration. If no value is provided by the user or by the dataTree object, parameters will be simulated according to default priors (see ?ellipseParam for information about default priors).}
}
\description{
Takes an object of class "dataTree", "detailedTree", or "phylo" and simulates values for the W, V, and XY fields, returning an object of class "dataTree"
}
\details{
This method assumes that a tree of class "detailedTree" or "phylo is provided, or the dataTree object already contains a tree of class "detailedTree" in the tree field. If you want to simulate a tree instead, use the sim_dataTree() function.
}
