% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTree.R
\name{sim_dataTree}
\alias{sim_dataTree}
\title{sim_dataTree function}
\usage{
sim_dataTree(param = NULL)
}
\arguments{
\item{param}{An object of type ellipseParam containing information about the model parameters associated with a simulation or MCMC iteration. If no value is provided by the user, parameters will be simulated according to default priors (see ?ellipseParam for information about default priors).}
}
\description{
Simulates a dataTree including the tree and all data fields
}
\details{
This method initializes a new dataTree and simulates using the sim() method
}
