% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailedTree.R
\name{sim_tree}
\alias{sim_tree}
\title{sim_tree() function}
\usage{
sim_tree()
}
\description{
Creates a new object of the "detailedTree" class using a simulated tree. May take a long time to generate for larger trees.
}
\details{
The simulated tree has height 1 and is generated using the package "phytools" through a pure-birth process with a rate of 4. Trees with less than 25 tips or more than 250 tips are rejected.
}
