% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_node.R
\name{transform_node}
\alias{transform_node}
\title{transform_node() function}
\usage{
transform_node(
  char = NULL,
  daughter = NULL,
  d = NULL,
  m = NULL,
  cval = NULL,
  hval = NULL,
  x = NULL,
  y = NULL,
  r = NULL,
  s = NULL,
  a = NULL,
  alpha = NULL,
  z = NULL
)
}
\arguments{
\item{char}{One of three continuous characters: "x", "y", or "a"}

\item{daughter}{One of two daughters: "left", "right"}

\item{d}{The daughter configuration during cladogenesis (0,1), where 0 makes the left daughter D1 and 1 makes the right daughter D1}

\item{m}{The cladogenetic mode (0,1), where 0 is budding and 1 is splitting}

\item{cval}{The c value (not index) of the cladogenetic scenario's concentric circle --> ex. 0.5}

\item{hval}{The h value (not index) of the cladogenetic scenario's direction line --> ex. pi/4, if required}

\item{x}{The x value to be transformed, if required}

\item{y}{The y value to be transformed, if required}

\item{r}{The data-augmented 'r' value associated with the cladogenesis event, if required}

\item{s}{The data-augmented 's' value associated with the cladogenesis event, if required}

\item{a}{The a value to be transformed, if required}

\item{alpha}{The 'alpha' model parameter, if required}

\item{z}{The 'z' model parameter, if required}
}
\description{
Accepts model parameters (alpha, z), information about a cladogenetic scenario (d, m, c, h), data-augmented values (r, s), and values for a character prior to cladogenesis (x, y, a) and deterministically calculates the new value of the character immediately after cladogenesis
}
